+++ product/src/main/java/br/com/venzel/product/modules/supplier/repositories/SupplierRepository.java
@@ -8,8 +8,8 @@ import org.springframework.stereotype.Repository;
 import br.com.venzel.product.modules.supplier.models.Supplier;
 
 @Repository
-public interface SupplierRepository extends JpaRepository<Supplier, Long> {
-    Optional<Supplier> findOneById(Long id);
+public interface SupplierRepository extends JpaRepository<Supplier, Integer> {
+    Optional<Supplier> findOneById(Integer id);
 
     Optional<Supplier> findOneByName(String name);
